requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
APScheduler==3.6.3
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==39.0.2
decorator==5.1.1
frozenlist==1.3.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
imageio==2.26.1
lazy_loader==0.1
magic-filter==1.0.9
multidict==6.0.4
networkx==3.0
numpy==1.24.2
packaging==23.0
Pillow==9.4.0
pycparser==2.21
python-dotenv==1.0.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
requests==2.28.2
rfc3986==1.5.0
scikit-image==0.20.0
scipy==1.9.1
six==1.16.0
sniffio==1.3.0
state==0.1.1.dev0
tifffile==2023.3.15
tornado==6.2
tzdata==2022.7
tzlocal==4.3
urllib3==1.26.15
yarl==1.8.2

